+++ service-proxy/src/main/java/no/mnemonic/services/common/api/proxy/client/ClientResultSet.java
@@ -5,10 +5,10 @@ import com.fasterxml.jackson.core.JsonToken;
 import lombok.CustomLog;
 import lombok.RequiredArgsConstructor;
 import no.mnemonic.commons.utilities.collections.SetUtils;
+import no.mnemonic.services.common.api.Resource;
 import no.mnemonic.services.common.api.ResultSet;
 import no.mnemonic.services.common.api.proxy.serializer.Serializer;
 
-import java.io.Closeable;
 import java.io.IOException;
 import java.util.Iterator;
 
@@ -23,7 +23,7 @@ public class ClientResultSet<T> implements ResultSet<T> {
   private final int limit;
   private final int offset;
   private final JsonParser data;
-  private final Closeable onClose;
+  private final Resource resource;
 
   private boolean closed;
 
@@ -76,13 +76,27 @@ public class ClientResultSet<T> implements ResultSet<T> {
   public void close() {
     if (closed) return;
     try {
+      if (resource != null) {
+        tryTo(resource::close, e -> LOGGER.error(e, "Error closing resource"));
+      }
       data.close();
       closed = true;
-      if (onClose != null) {
-        tryTo(onClose::close, e -> LOGGER.error(e, "Error closing resource"));
-      }
     } catch (IOException e) {
       throw new IllegalStateException("Error closing stream", e);
+    }
+  }
+
+  @Override
+  public void cancel() {
+    if (closed) return;
+    try {
+      if (resource != null) {
+        tryTo(resource::cancel, e -> LOGGER.error(e, "Error cancelling resource"));
+      }
+      data.close();
+      closed = true;
+    } catch (IOException e) {
+      throw new IllegalStateException("Error cancelling stream", e);
     }
   }